fix: Move json-schema to static folder to be served correctly
